# fix concurrent map write issue by init all regexp when startup