this p-r is to fix https://github.com/gorhill/cronexpr/issues/19

please help to review it when you get a chance. thx~

i put layoutRegexpMap to fieldDescriptor, and use fieldDescriptor's pointer.
ut passed. 
run benchmark, no obvious performance lose. 
BenchMark Before:  
BenchmarkParse-8   	  100000	     16959 ns/op	    4553 B/op	      65 allocs/op
BenchmarkNext-8    	  300000	      4873 ns/op	     453 B/op	      19 allocs/op

After:
BenchmarkParse-8   	  100000	     17210 ns/op	    4553 B/op	      65 allocs/op
BenchmarkNext-8    	  300000	      4890 ns/op	     453 B/op	      19 allocs/op