@aiquestion 2017-01-07
@elwinar as golang doc : https://golang.org/ref/mem.
and this blog: http://blog.launchdarkly.com/golang-pearl-thread-safe-writes-and-double-checked-locking-in-go/
i need to use sync/atomic to load and set checked value in double checked locking. So i choose the suggested way once.Do().  :-) this function is a correct implementation of double checked locking :-)

but actually my previous p-r seems to have another problem: if a goroutine is setting the value, while another goroutine tries to read and check the value, a panic will still occurs. golang map doesn't not allow read and write at the same time.