@elwinar 2017-01-07
What I meant was that you only locked once, so it couldn't be a 
double-checked lock. Hence the warning in the post you linked: without 
the read lock, the go memory model doesn't guarantee you that you're 
thread-safe.

Examples are always better. This (from your PR) is not a double-checked 
lock:

```
func makeLayoutRegexp(layout, value string) *regexp.Regexp {
	layout = strings.Replace(layout, `%value%`, value, -1)
	re := layoutRegexp[layout]
	if re == nil {
		// double check locking to fix issue: 
https://github.com/gorhill/cronexpr/issues/19
		layoutRegexpLock.Lock()
		defer layoutRegexpLock.Unlock()
		re = layoutRegexp[layout]
		if re == nil {
			re = regexp.MustCompile(layout)
			layoutRegexp[layout] = re
		}
	}
	return re
}
```

This is one:

```
func makeLayoutRegexp(layout, value string) *regexp.Regexp {
	layout = strings.Replace(layout, `%value%`, value, -1)
	layoutRegexpLock.RLock()
	re := layoutRegexp[layout]
	if re == nil {
		layoutRegexpLock.RUnlock()
		// double check locking to fix issue: 
https://github.com/gorhill/cronexpr/issues/19
		layoutRegexpLock.Lock()
		defer layoutRegexpLock.Unlock()
		re = layoutRegexp[layout]
		if re == nil {
			re = regexp.MustCompile(layout)
			layoutRegexp[layout] = re
		}
	} else {
		layoutRegexpLock.RUnlock()
	}
	return re
}
```