@aiquestion 2017-01-10
ah. you are right.
I think i wrote a java-pattern double checked locking in my previous p-r which is wrong in golang.